refactor: refactor js according es-6
