+++ src/index.js
@@ -1,4 +1,21 @@
 import './sass/styles.scss';
+import { createKeyboardRow, renderKeyboardKey } from './utils/utils';
+import {
+  firstRowKeys,
+  firstRowKeysCodes,
+  secondRowKeysCodes,
+  secondRowKeys,
+  thirdRowKeysCodes,
+  thirdRowKeys,
+  fourthRowKeysCodes,
+  fourthRowKeys,
+  fifthRowKeysCodes,
+  fifthRowKeys,
+  capsModeKeys,
+  shiftModeKeys,
+  shiftModeKeysActiveValues,
+  shiftModeKeysInactiveValues,
+} from './utils/constants';
 
 let keyboardSettings;
 
@@ -22,356 +39,64 @@ const keyboard = document.createElement('div');
 keyboard.classList.add('keyboard');
 body.append(keyboard);
 
-/* First Row */
-
-const keyboardFirstRow = document.createElement('div');
-keyboardFirstRow.classList.add('keyboard-row');
-keyboard.append(keyboardFirstRow);
-
-const firstRowKeysCodes = [
-  'Backquote',
-  'Digit1',
-  'Digit2',
-  'Digit3',
-  'Digit4',
-  'Digit5',
-  'Digit6',
-  'Digit7',
-  'Digit8',
-  'Digit9',
-  'Digit0',
-  'Minus',
-  'Equal',
-  'Backspace',
-];
-
-const firstRowKeys = [
-  '`',
-  '1',
-  '2',
-  '3',
-  '4',
-  '5',
-  '6',
-  '7',
-  '8',
-  '9',
-  '0',
-  '-',
-  '=',
-  'Backspace',
-];
+const shiftAndCapsMode = {};
 
 const copyedFirstRowKeys = [...firstRowKeys];
+const copyedSecondRowKeys = [...secondRowKeys];
+const copyedThirdRowKeys = [...thirdRowKeys];
+const copyedFourthRowKeys = [...fourthRowKeys];
+const copyedFifthRowKeys = [...fifthRowKeys];
 
+/* First Row */
+const keyboardFirstRow = createKeyboardRow();
+firstRowKeys.forEach((key) => {
+  keyboardFirstRow.append(
+    renderKeyboardKey(key, copyedFirstRowKeys, firstRowKeysCodes, keyboardSettings, capsModeKeys),
+  );
+});
 /* ============================================== */
 
 /* Second Row */
-const keyboardSecondRow = document.createElement('div');
-keyboardSecondRow.classList.add('keyboard-row');
-keyboard.append(keyboardSecondRow);
-
-const secondRowKeysCodes = [
-  'Tab',
-  'KeyQ',
-  'KeyW',
-  'KeyE',
-  'KeyR',
-  'KeyT',
-  'KeyY',
-  'KeyU',
-  'KeyI',
-  'KeyO',
-  'KeyP',
-  'BracketLeft',
-  'BracketRight',
-  'Backslash',
-  'Delete',
-];
-
-const secondRowKeys = [
-  'Tab',
-  'q',
-  'w',
-  'e',
-  'r',
-  't',
-  'y',
-  'u',
-  'i',
-  'o',
-  'p',
-  '[',
-  ']',
-  '\\',
-  'Delete',
-];
-
-const copyedSecondRowKeys = [...secondRowKeys];
-
+const keyboardSecondRow = createKeyboardRow();
+secondRowKeys.forEach((key) => {
+  keyboardSecondRow.append(
+    renderKeyboardKey(key, copyedSecondRowKeys, secondRowKeysCodes, keyboardSettings, capsModeKeys),
+  );
+});
 /* ============================================== */
 
 /* Third Row */
-const keyboardThirdRow = document.createElement('div');
-keyboardThirdRow.classList.add('keyboard-row');
-keyboard.append(keyboardThirdRow);
-
-const thirdRowKeysCodes = [
-  'CapsLock',
-  'KeyA',
-  'KeyS',
-  'KeyD',
-  'KeyF',
-  'KeyG',
-  'KeyH',
-  'KeyJ',
-  'KeyK',
-  'KeyL',
-  'Semicolon',
-  'Quote',
-  'Enter',
-];
-
-const thirdRowKeys = [
-  'CapsLock',
-  'a',
-  's',
-  'd',
-  'f',
-  'g',
-  'h',
-  'j',
-  'k',
-  'l',
-  ';',
-  "'",
-  'Enter',
-];
-
-const copyedThirdRowKeys = [...thirdRowKeys];
-
+const keyboardThirdRow = createKeyboardRow();
+thirdRowKeys.forEach((key) => {
+  keyboardThirdRow.append(
+    renderKeyboardKey(key, copyedThirdRowKeys, thirdRowKeysCodes, keyboardSettings, capsModeKeys),
+  );
+});
 /* ============================================== */
 
 /* Fourth Row */
-const keyboardFourthRow = document.createElement('div');
-keyboardFourthRow.classList.add('keyboard-row');
-keyboard.append(keyboardFourthRow);
-
-const fourthRowKeysCodes = [
-  'ShiftLeft',
-  'KeyZ',
-  'KeyX',
-  'KeyC',
-  'KeyV',
-  'KeyB',
-  'KeyN',
-  'KeyM',
-  'Comma',
-  'Period',
-  'Slash',
-  'ArrowUp',
-  'ShiftRight',
-];
-
-const fourthRowKeys = [
-  'Shift',
-  'z',
-  'x',
-  'c',
-  'v',
-  'b',
-  'n',
-  'm',
-  ',',
-  '.',
-  '/',
-  'ArrowUp',
-  'Shift',
-];
-
-const copyedFourthRowKeys = [...fourthRowKeys];
-
+const keyboardFourthRow = createKeyboardRow();
+fourthRowKeys.forEach((key) => {
+  keyboardFourthRow.append(
+    renderKeyboardKey(key, copyedFourthRowKeys, fourthRowKeysCodes, keyboardSettings, capsModeKeys),
+  );
+});
 /* ============================================== */
 
 /* Fifth Row */
-
-const keyboardFifthRow = document.createElement('div');
-keyboardFifthRow.classList.add('keyboard-row');
-keyboard.append(keyboardFifthRow);
-
-const fifthRowKeysCodes = [
-  'ControlLeft',
-  'AltLeft',
-  'Space',
-  'AltRight',
-  'ArrowLeft',
-  'ArrowDown',
-  'ArrowRight',
-  'ControlRight',
-];
-
-const fifthRowKeys = [
-  'Control',
-  'Alt',
-  ' ',
-  'Alt',
-  'ArrowLeft',
-  'ArrowDown',
-  'ArrowRight',
-  'Control',
-];
-
-const copyedFifthRowKeys = [...fifthRowKeys];
-
-/* ============================================== */
-
-/* CapsMode Keys */
-
-const capsModeKeys = [
-  'KeyQ',
-  'KeyW',
-  'KeyE',
-  'KeyR',
-  'KeyT',
-  'KeyY',
-  'KeyU',
-  'KeyI',
-  'KeyO',
-  'KeyP',
-  'KeyA',
-  'KeyS',
-  'KeyD',
-  'KeyF',
-  'KeyG',
-  'KeyH',
-  'KeyJ',
-  'KeyK',
-  'KeyL',
-  'KeyZ',
-  'KeyX',
-  'KeyC',
-  'KeyV',
-  'KeyB',
-  'KeyN',
-  'KeyM',
-];
-
-/* ============================================== */
-
-/* ShiftMode Keys */
-
-const shiftModeKeys = [
-  'Backquote',
-  'Digit1',
-  'Digit2',
-  'Digit3',
-  'Digit4',
-  'Digit5',
-  'Digit6',
-  'Digit7',
-  'Digit8',
-  'Digit9',
-  'Digit0',
-  'Minus',
-  'Equal',
-  'BracketLeft',
-  'BracketRight',
-  'Backslash',
-  'Semicolon',
-  'Quote',
-  'Comma',
-  'Period',
-  'Slash',
-];
-
-const shiftModeKeysActiveValues = [
-  '~',
-  '!',
-  '@',
-  '#',
-  '$',
-  '%',
-  '^',
-  '&',
-  '*',
-  '(',
-  ')',
-  '_',
-  '+',
-  '{',
-  '}',
-  '|',
-  ':',
-  '"',
-  '<',
-  '>',
-  '?',
-];
-
-const shiftModeKeysInactiveValues = [
-  '`',
-  '1',
-  '2',
-  '3',
-  '4',
-  '5',
-  '6',
-  '7',
-  '8',
-  '9',
-  '0',
-  '-',
-  '=',
-  '[',
-  ']',
-  '\\',
-  ';',
-  "'",
-  ',',
-  '.',
-  '/',
-];
-
-/* ============================================== */
-
-const renderKeyboardKey = (key, rowKeys, rowKeysCodes) => {
-  const keyboardKey = document.createElement('button');
-  keyboardKey.classList.add('keyboard-key');
-  if (key === 'Backspace' || key === 'Tab' || key === 'CapsLock' || key === 'Enter' || key === 'Shift' || key === ' ') keyboardKey.classList.add('flex-grow');
-  if (key === 'CapsLock' && keyboardSettings.capsMode) keyboardKey.classList.add('power-on');
-  keyboardKey.dataset.key = rowKeysCodes[rowKeys.indexOf(key)];
-  keyboardKey.textContent = key;
-  if (key === 'ArrowUp') keyboardKey.textContent = '↑';
-  if (key === 'ArrowLeft') keyboardKey.textContent = '←';
-  if (key === 'ArrowDown') keyboardKey.textContent = '↓';
-  if (key === 'ArrowRight') keyboardKey.textContent = '→';
-  if (capsModeKeys.some((keyCode) => rowKeysCodes[rowKeys.indexOf(key)] === keyCode)
-    && keyboardSettings.capsMode) keyboardKey.textContent = key.toUpperCase();
-  rowKeys.splice(rowKeys.indexOf(key), 1, '');
-  return keyboardKey;
-};
-
-firstRowKeys.forEach((key) => {
-  keyboardFirstRow.append(renderKeyboardKey(key, copyedFirstRowKeys, firstRowKeysCodes));
-});
-
-secondRowKeys.forEach((key) => {
-  keyboardSecondRow.append(renderKeyboardKey(key, copyedSecondRowKeys, secondRowKeysCodes));
-});
-
-thirdRowKeys.forEach((key) => {
-  keyboardThirdRow.append(renderKeyboardKey(key, copyedThirdRowKeys, thirdRowKeysCodes));
-});
-
-fourthRowKeys.forEach((key) => {
-  keyboardFourthRow.append(renderKeyboardKey(key, copyedFourthRowKeys, fourthRowKeysCodes));
-});
-
+const keyboardFifthRow = createKeyboardRow();
 fifthRowKeys.forEach((key) => {
-  keyboardFifthRow.append(renderKeyboardKey(key, copyedFifthRowKeys, fifthRowKeysCodes));
+  keyboardFifthRow.append(
+    renderKeyboardKey(key, copyedFifthRowKeys, fifthRowKeysCodes, keyboardSettings, capsModeKeys),
+  );
 });
+/* ============================================== */
 
-const shiftAndCapsMode = {};
+keyboard.append(keyboardFirstRow);
+keyboard.append(keyboardSecondRow);
+keyboard.append(keyboardThirdRow);
+keyboard.append(keyboardFourthRow);
+keyboard.append(keyboardFifthRow);
 
 document.addEventListener('keydown', (e) => {
   document.querySelector('.output').focus();
@@ -479,8 +204,3 @@ document.addEventListener('keyup', (e) => {
     });
   }
 });
-
-// document.addEventListener('keydown', (e) => {
-//   shiftModeKeysInactiveValues.push(e.key);
-//   console.log(shiftModeKeysInactiveValues);
-// });
